+++ etna/loggers/base.py
@@ -5,6 +5,7 @@ from typing import Any
 from typing import Dict
 from typing import Union
 
+import numpy as np
 import pandas as pd
 
 from etna.core.mixins import BaseMixin
@@ -132,7 +133,7 @@ class _Logger(BaseLogger):
         ts:
             TSDataset to with backtest data
         metrics_df:
-            Dataframe produced with TimeSeriesCrossValidation.get_metrics(aggregate_metrics=False)
+            Dataframe produced with Pipeline._get_backtest_metrics()
         forecast_df:
             Forecast from backtest
         fold_info_df:
@@ -143,7 +144,7 @@ class _Logger(BaseLogger):
 
     def log_backtest_run(self, metrics: pd.DataFrame, forecast: pd.DataFrame, test: pd.DataFrame):
         """
-        Backtest metrics to logger.
+        Backtest metrics from one fold to logger.
 
         Parameters
         ----------
@@ -171,3 +172,49 @@ class _Logger(BaseLogger):
     def pl_loggers(self):
         """Pytorch lightning loggers."""
         return [logger.pl_logger for logger in self.loggers if "_pl_logger" in vars(logger)]
+
+
+def percentile(n: int):
+    """Percentile for pandas agg."""
+
+    def percentile_(x):
+        return np.percentile(x.values, n)
+
+    percentile_.__name__ = "percentile_%s" % n
+    return percentile_
+
+
+def aggregate_metrics_df(metrics_df: pd.DataFrame) -> Dict[str, float]:
+    """Aggregate metrics in `log_backtest_metrics` method.
+
+    Parameters
+    ----------
+    metrics_df:
+        Dataframe produced with Pipeline._get_backtest_metrics()
+    """
+    # case for aggregate_metrics=False
+    if "fold_number" in metrics_df.columns:
+        metrics_dict = (
+            metrics_df.groupby("segment")
+            .mean()
+            .reset_index()
+            .drop(["segment", "fold_number"], axis=1)
+            .apply(["median", "mean", "std", percentile(5), percentile(25), percentile(75), percentile(95)])
+            .to_dict()
+        )
+
+    # case for aggregate_metrics=True
+    else:
+        metrics_dict = (
+            metrics_df.drop(["segment"], axis=1)
+            .apply(["median", "mean", "std", percentile(5), percentile(25), percentile(75), percentile(95)])
+            .to_dict()
+        )
+
+    metrics_dict_wide = {
+        f"{metrics_key}_{statistics_key}": value
+        for metrics_key, values in metrics_dict.items()
+        for statistics_key, value in values.items()
+    }
+
+    return metrics_dict_wide