+++ etna/loggers/wandb_logger.py
@@ -7,11 +7,11 @@ from typing import Optional
 from typing import Union
 from uuid import uuid4
 
-import numpy as np
 import pandas as pd
 
 from etna import SETTINGS
 from etna.loggers.base import BaseLogger
+from etna.loggers.base import aggregate_metrics_df
 
 if TYPE_CHECKING:
     from pytorch_lightning.loggers import WandbLogger as PLWandbLogger
@@ -20,16 +20,6 @@ if TYPE_CHECKING:
 
 if SETTINGS.wandb_required:
     import wandb
-
-
-def percentile(n: int):
-    """Percentile for pandas agg."""
-
-    def percentile_(x):
-        return np.percentile(x.values, n)
-
-    percentile_.__name__ = "percentile_%s" % n
-    return percentile_
 
 
 class WandbLogger(BaseLogger):
@@ -46,7 +36,7 @@ class WandbLogger(BaseLogger):
         plot: bool = True,
         table: bool = True,
         name_prefix: str = "",
-        config: Optional[Union[Dict, str, None]] = None,
+        config: Optional[Dict[str, Any]] = None,
         log_model: bool = False,
     ):
         """
@@ -124,8 +114,10 @@ class WandbLogger(BaseLogger):
 
         Parameters
         ----------
+        ts:
+            TSDataset to with backtest data
         metrics_df:
-            Dataframe produced with Pipeline._get_backtest_metrics() or TimeSeriesCrossValidation.get_metrics()
+            Dataframe produced with Pipeline._get_backtest_metrics()
         forecast_df:
             Forecast from backtest
         fold_info_df:
@@ -143,26 +135,9 @@ class WandbLogger(BaseLogger):
             fig = plot_backtest_interactive(forecast_df, ts, history_len=100)
             self.experiment.log({"backtest": fig})
 
-        # case for aggregate_metrics=False
-        if "fold_number" in metrics_df.columns:
-            metrics_dict = (
-                metrics_df.groupby("segment")
-                .mean()
-                .reset_index()
-                .drop(["segment", "fold_number"], axis=1)
-                .apply(["median", "mean", "std", percentile(5), percentile(25), percentile(75), percentile(95)])
-                .to_dict()
-            )
-        # case for aggregate_metrics=True
-        else:
-            metrics_dict = (
-                metrics_df.drop(["segment"], axis=1)
-                .apply(["median", "mean", "std", percentile(5), percentile(25), percentile(75), percentile(95)])
-                .to_dict()
-            )
-        for metrics_key, values in metrics_dict.items():
-            for statistics_key, value in values.items():
-                self.experiment.summary[f"{metrics_key}_{statistics_key}"] = value
+        metrics_dict = aggregate_metrics_df(metrics_df)
+        for metric_key, metric_value in metrics_dict.items():
+            self.experiment.summary[metric_key] = metric_value
 
     def log_backtest_run(self, metrics: pd.DataFrame, forecast: pd.DataFrame, test: pd.DataFrame):
         """
@@ -180,21 +155,16 @@ class WandbLogger(BaseLogger):
         from etna.datasets import TSDataset
 
         columns_name = list(metrics.columns)
-        metrics.reset_index(inplace=True)
+        metrics = metrics.reset_index()
         metrics.columns = ["segment"] + columns_name
         if self.table:
             self.experiment.summary["metrics"] = wandb.Table(data=metrics)
             self.experiment.summary["forecast"] = wandb.Table(data=TSDataset.to_flatten(forecast))
             self.experiment.summary["test"] = wandb.Table(data=TSDataset.to_flatten(test))
 
-        metrics_dict = (
-            metrics.drop(["segment"], axis=1)
-            .apply(["median", "mean", "std", percentile(5), percentile(25), percentile(75), percentile(95)])
-            .to_dict()
-        )
-        for metrics_key, values in metrics_dict.items():
-            for statistics_key, value in values.items():
-                self.experiment.summary[f"{metrics_key}_{statistics_key}"] = value
+        metrics_dict = aggregate_metrics_df(metrics)
+        for metric_key, metric_value in metrics_dict.items():
+            self.experiment.summary[metric_key] = metric_value
 
     def start_experiment(self, job_type: Optional[str] = None, group: Optional[str] = None, *args, **kwargs):
         """Start experiment(logger post init or reinit next experiment with the same name).