+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Catboost version too high ([#394](https://github.com/tinkoff-ai/etna/pull/394))
 - Add sorting of classes in left bar in docs ([#397](https://github.com/tinkoff-ai/etna/pull/397))
 - nn notebook in docs ([#396](https://github.com/tinkoff-ai/etna/pull/396))
+- SklearnTransform column name generation ([#398](https://github.com/tinkoff-ai/etna/pull/398))
 
 ## [1.4.2] - 2021-12-09
 ### Fixed